[tf.data] graduate experimental take_while api to tf.data.Dataset
