+++ tensorflow/python/data/experimental/ops/take_while_ops.py
@@ -17,41 +17,11 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from tensorflow.python.data.ops import dataset_ops
-from tensorflow.python.framework import dtypes
-from tensorflow.python.framework import tensor_spec
-from tensorflow.python.ops import gen_experimental_dataset_ops
+from tensorflow.python.util import deprecation
 from tensorflow.python.util.tf_export import tf_export
 
 
-class _TakeWhileDataset(dataset_ops.UnaryUnchangedStructureDataset):
-  """A dataset that stops iteration when `predicate` returns false."""
-
-  def __init__(self, input_dataset, predicate):
-    """See `take_while()` for details."""
-
-    self._input_dataset = input_dataset
-    wrapped_func = dataset_ops.StructuredFunctionWrapper(
-        predicate,
-        "tf.data.experimental.take_while()",
-        dataset=self._input_dataset)
-
-    if not wrapped_func.output_structure.is_compatible_with(
-        tensor_spec.TensorSpec([], dtypes.bool)):
-      raise ValueError("`predicate` must return a scalar boolean tensor.")
-
-    self._predicate = wrapped_func
-    var_tensor = gen_experimental_dataset_ops.take_while_dataset(
-        self._input_dataset._variant_tensor,  # pylint: disable=protected-access
-        other_arguments=self._predicate.function.captured_inputs,
-        predicate=self._predicate.function,
-        **self._flat_structure)
-    super(_TakeWhileDataset, self).__init__(input_dataset, var_tensor)
-
-  def _functions(self):
-    return [self._predicate]
-
-
+@deprecation.deprecated(None, "Use `tf.data.Dataset.take_while(...)")
 @tf_export("data.experimental.take_while")
 def take_while(predicate):
   """A transformation that stops dataset iteration based on a `predicate`.
@@ -67,6 +37,6 @@ def take_while(predicate):
   """
 
   def _apply_fn(dataset):
-    return _TakeWhileDataset(dataset, predicate)
+    return dataset.take_while(predicate=predicate)
 
   return _apply_fn
